fix(box-utils): Renamme canvas-utils to box-utils
